fix: MeanFlow review findings — x_r_pred rename, explicit interval dt, combined_hidden dtype, r-sampling config
- Rename MeanFlow return key from x_0_pred to x_r_pred (semantically correct when r>0)
- Use explicit (t_i - r_i) instead of scalar dt for robustness to non-uniform timesteps
- Add dtype=xxx.dtype to combined_hidden in training and inference to avoid implicit float32
- Add meanflow_r_zero_prob config (default 0.5) for ablating r=0 sampling ratio
- Keep autocast disabled in JVP:实测 bf16 JVP has ~16% error on full model
- Fix verbose step counter to print 1→N (not N→1) for clarity
- Add v_target.float() in loss for consistent fp32 MSE
- Add 3 new tests: config r_zero_prob, x_r_pred key, r_zero_prob custom
- 54/54 tests pass

+++ src/mobius/translation/modeling_meanflow.py
@@ -157,7 +157,8 @@ class MeanFlowHead(nn.Module):
         Returns:
             (loss, u) — loss is differentiable, u is the primal prediction.
         """
-        # Disable autocast for JVP precision (bf16 tangents lose precision)
+        # Disable autocast for JVP precision (bf16 tangents lose ~16% relative
+        # error on full model; FP32 JVP is required for correct du/dt).
         with torch.amp.autocast(device_type=t.device.type, enabled=False):
             u, du_dt = jvp(pipeline_fn, (t.float(),), (torch.ones_like(t).float(),))
 
@@ -165,7 +166,7 @@ class MeanFlowHead(nn.Module):
         dt = (t - r)
         V_theta = u + dt.view(-1, 1, 1) * du_dt.detach()
 
-        loss = F.mse_loss(V_theta, v_target)
+        loss = F.mse_loss(V_theta, v_target.float())
         return loss, u
 
     def compute_uloss(
@@ -195,13 +196,14 @@ class MeanFlowHead(nn.Module):
         Returns:
             (loss, u) — loss 可微，u 为原始预测。
         """
-        # Disable autocast for JVP precision (bf16 tangents lose precision)
+        # Disable autocast for JVP precision (bf16 tangents lose ~16% relative
+        # error on full model; FP32 JVP is required for correct du/dt).
         with torch.amp.autocast(device_type=t.device.type, enabled=False):
             u, du_dt = jvp(pipeline_fn, (t.float(),), (torch.ones_like(t).float(),))
 
         # u-loss: target = v_gt - (t - r) * du/dt, 无 detach
         dt = (t - r)
-        u_target = v_target - dt.view(-1, 1, 1) * du_dt
+        u_target = v_target.float() - dt.view(-1, 1, 1) * du_dt
 
         loss = F.mse_loss(u, u_target)
         return loss, u
